break down protocol module
